+++ db/collection.go
@@ -27,7 +27,6 @@ import (
 	"github.com/lens-vm/lens/host-go/config/model"
 	"github.com/sourcenetwork/immutable"
 
-	"github.com/sourcenetwork/defradb/acp"
 	"github.com/sourcenetwork/defradb/client"
 	"github.com/sourcenetwork/defradb/client/request"
 	"github.com/sourcenetwork/defradb/core"
@@ -36,6 +35,7 @@ import (
 	"github.com/sourcenetwork/defradb/db/fetcher"
 	"github.com/sourcenetwork/defradb/errors"
 	"github.com/sourcenetwork/defradb/events"
+	"github.com/sourcenetwork/defradb/internal/acp"
 	"github.com/sourcenetwork/defradb/lens"
 	merklecrdt "github.com/sourcenetwork/defradb/merkle/crdt"
 )
