+++ db/collection_delete.go
@@ -13,10 +13,10 @@ package db
 import (
 	"context"
 
-	"github.com/sourcenetwork/defradb/acp"
 	"github.com/sourcenetwork/defradb/client"
 	"github.com/sourcenetwork/defradb/core"
 	"github.com/sourcenetwork/defradb/events"
+	"github.com/sourcenetwork/defradb/internal/acp"
 	"github.com/sourcenetwork/defradb/merkle/clock"
 )
